T2-54 #done All methods provided to store an Exercise with images to the database. Exercises can be retrieved from database with images paths. Image upload has some restrictions given to the mod at mod_create_exercise. To see the uploaded pictures without server restart, look in the pom.xml at <build> for instructions.

+++ src/main/java/de/ep/team2/core/service/ExerciseService.java
@@ -1,6 +1,8 @@
 package de.ep.team2.core.service;
 
 import de.ep.team2.core.entities.Exercise;
+import de.ep.team2.core.enums.ImageType;
+import de.ep.team2.core.enums.WeightType;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -8,6 +10,10 @@ import java.nio.file.*;
 import java.util.LinkedList;
 import java.util.List;
 
+/**
+ * Forwards requests to the Database Service and manages the Storage of the
+ * Exercise images. Tested with Acceptance Test.
+ */
 public class ExerciseService {
 
     public Exercise getExerciseById(int id) {
@@ -25,25 +31,29 @@ public class ExerciseService {
      * type of image the file is.
      *
      * @param name name of the exercise to insert.
-     * @param description description of the exercise to insert.
-     * @param muscleImgPaths Paths to the muscle images.
-     * @param otherImgPaths Paths to the other images.
+     * @param description optional description of the exercise to insert.
+     * @param weightType weight type of the exercise to insert.
+     * @param link optional link to a example video of the exercise.
+     * @param muscleImgPaths optional Paths to the muscle images.
+     * @param otherImgPaths optional Paths to the other images.
      * @return The id of the just inserted exercise.
      */
     public Integer insertExercise(String name, String description,
+                                  WeightType weightType, String link,
                                   List<String> muscleImgPaths, List<String> otherImgPaths) {
         LinkedList<String[]> imgPaths = new LinkedList<>();
         if (muscleImgPaths != null) {
             for (String s : muscleImgPaths) {
-                imgPaths.add(new String[]{s, "muscle"});
+                imgPaths.add(new String[]{s, ImageType.MUSCLE_IMAGE.toString()});
             }
         }
         if (otherImgPaths != null) {
             for (String s : otherImgPaths) {
-                imgPaths.add(new String[]{s,"other"});
+                imgPaths.add(new String[]{s, ImageType.OTHER_IMAGE.toString()});
             }
         }
-        return DataBaseService.getInstance().insertExercise(name, description, imgPaths);
+        return DataBaseService.getInstance().insertExercise(name, description,
+                weightType, link, imgPaths);
     }
 
     /**
@@ -53,6 +63,7 @@ public class ExerciseService {
      * @param image image to save.
      * @param exercise exercise of the image.
      * @return The relative path, used by thymeleaf ,where the image is saved.
+     *          null when an errors appears.
      */
     public String uploadImg(MultipartFile image, Exercise exercise) {
         String fileName = image.getOriginalFilename();
@@ -82,13 +93,17 @@ public class ExerciseService {
         String dirDestString = "src/main/resources/static/images/" +  toDelete.getName();
         Path dirDest = Paths.get(dirDestString);
         try {
-            for (String s : toDelete.getOtherImgPaths()) {
-                Path dest = Paths.get("src/main/resources/static" + s);
-                Files.delete(dest);
+            if (!toDelete.getOtherImgPaths().isEmpty()) {
+                for (String s : toDelete.getOtherImgPaths()) {
+                    Path dest = Paths.get("src/main/resources/static" + s);
+                    Files.delete(dest);
+                }
             }
-            for (String s : toDelete.getMuscleImgPaths()) {
-                Path dest = Paths.get("src/main/resources/static" + s);
-                Files.delete(dest);
+            if (!toDelete.getMuscleImgPaths().isEmpty()) {
+                for (String s : toDelete.getMuscleImgPaths()) {
+                    Path dest = Paths.get("src/main/resources/static" + s);
+                    Files.delete(dest);
+                }
             }
             Files.delete(dirDest);
         } catch (IOException exception) {
